# src/public/swagger.json
{
  "swagger":"2.0",
  "info":{
    "description":"This is a restful API for access to datahub investment project",
    "version":"2.0.0",
    "title":"Datahub API",
    "contact":{
      "email":"zac@thetolleys.com"
    },
    "license":{
      "name":"MIT"
    }
  },
  "tags":[
    {
      "name":"admins",
      "description":"Secured Admin-only calls"
    },
    {
      "name":"user",
      "description":"Operations for authenticated users"
    }
  ],
  "schemes":[
    "https"
  ],
  "paths":{
    "/company/":{
      "post":{
        "summary": "Create new company",
        "description":"Creates a new company, pass the company object in the body",
        "operationId":"addCompany",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"company",
            "in":"body",
            "description":"compay to add",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Company"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"company response",
            "schema":{
              "$ref":"#/definitions/Company"
            }
          }
        }
      }
    },
    "/company/{id}/":{
      "get":{
        "summary":"Retrieve company data",
        "operationId":"getCompany",
        "description":"Pass the ID of the company to retrieve its details",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"path",
            "name":"id",
            "description":"The unique UUID for the company",
            "required":true,
            "type":"string"
          }
        ],
        "responses":{
          "200":{
            "description":"company for id",
            "schema":{
              "$ref":"#/definitions/Company"
            }
          }
        }
      },
      "put":{
        "summary": "Update an existing company",
        "description":"Update an existing company",
        "operationId":"updateCompany",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"path",
            "name":"id",
            "description":"The unique UUID for the company",
            "required":true,
            "type":"string"
          },
          {
            "name":"company",
            "in":"body",
            "description":"compay to add",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Company"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"company response",
            "schema":{
              "$ref":"#/definitions/Company"
            }
          }
        }
      }
    },
    "/company/{id}/contacts/":{
      "get":{
        "summary":"Get a list of contacts for a company",
        "operationId":"companycontacts",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"path",
            "name":"id",
            "description":"The unique UUID for the company",
            "required":true,
            "type":"string"
          }
        ],
        "responses":{
          "200":{
            "description":"Contact list",
            "schema":{
              "type":"array",
              "items":{
                "$ref":"#/definitions/Contact"
              }
            }
          }
        }
      }
    },
    "/company/{id}/investmentprojects/":{
      "get":{
        "summary":"Get a list of investment projects for a company",
        "operationId":"companyproject",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"path",
            "name":"id",
            "description":"The unique UUID for the company",
            "required":true,
            "type":"string"
          }
        ],
        "responses":{
          "200":{
            "description":"Investment project list",
            "schema":{
              "type":"array",
              "items":{
                "$ref":"#/definitions/InvestmentProject"
              }
            }
          }
        }
      }
    },
    "/ch-company/{company_number}/":{
      "get":{
        "summary":"Retrieve companies house data",
        "operationId":"getCompaniesHouse",
        "description":"Pass the company number of the company to retrieve its details",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"path",
            "name":"company_number",
            "description":"The companies house company number",
            "required":true,
            "type":"string"
          }
        ],
        "responses":{
          "200":{
            "description":"company for company number",
            "schema":{
              "$ref":"#/definitions/CompaniesHouseCompany"
            }
          }
        }
      }
    },
    "/contact/":{
      "post":{
        "summary":"Creates a new contact",
        "operationId":"addcontact",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"contact",
            "in":"body",
            "description":"contact to add",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Contact"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"Contact response",
            "schema":{
              "$ref":"#/definitions/Contact"
            }
          }
        }
      }
    },
    "/contact/{id}/":{
      "get":{
        "summary":"Retrieve contact data",
        "operationId":"getcontact",
        "description":"Pass the ID of the contact to retrieve its details",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"path",
            "name":"id",
            "description":"The unique UUID for the contact",
            "required":true,
            "type":"string"
          }
        ],
        "responses":{
          "200":{
            "description":"Contact for id",
            "schema":{
              "$ref":"#/definitions/Contact"
            }
          }
        }
      },
      "put":{
        "summary":"Update a new contact",
        "operationId":"updatecontact",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"path",
            "name":"id",
            "description":"The unique UUID for the contact",
            "required":true,
            "type":"string"
          },
          {
            "name":"contact",
            "in":"body",
            "description":"Contact to add",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Contact"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"Contact response",
            "schema":{
              "$ref":"#/definitions/Contact"
            }
          }
        }
      }
    },
    "/metadata/{table}/":{
      "get":{
        "summary":"Get a list values for a metadata list",
        "operationId":"metadatalist",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"path",
            "name":"table",
            "description":"The name of the metadata entity to list",
            "required":true,
            "type":"string"
          }
        ],
        "responses":{
          "200":{
            "description":"Value list",
            "schema":{
              "type":"array",
              "items":{
                "$ref":"#/definitions/Metadata"
              }
            }
          }
        }
      }
    },
    "/metadata/{table}/{id}/":{
      "get":{
        "summary":"Retrieve company data",
        "operationId":"getmetadatatem",
        "description":"Pass the ID of the company to retrieve its details",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"path",
            "name":"table",
            "description":"name of the related table to look in",
            "required":true,
            "type":"string"
          },
          {
            "in":"path",
            "name":"id",
            "description":"The unique UUID for the related data",
            "required":true,
            "type":"string"
          }
        ],
        "responses":{
          "200":{
            "description":"data for id",
            "schema":{
              "$ref":"#/definitions/Metadata"
            }
          }
        }
      }
    },
    "/search/":{
      "get":{
        "summary":"Search for companies",
        "operationId":"search",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"term",
            "description":"The string to search for",
            "required":true,
            "type":"string"
          }
        ],
        "responses":{
          "200":{
            "description":"Search Results",
            "schema":{
              "$ref":"#/definitions/SearchResults"
            }
          }
        }
      }
    }
  },
  "definitions":{
    "Company":{
      "type":"object",
      "required":[
        "id",
        "name"
      ],
      "properties":{
        "id":{
          "type":"string",
          "format":"uuid",
          "example":"d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "name":{
          "type":"string",
          "example":"Widgets r us"
        },
        "registered_address_1":{
          "type":"string",
          "example":"1 Smith Street"
        },
        "registered_address_2":{
          "type":"string"
        },
        "registered_address_3":{
          "type":"string"
        },
        "registered_address_4":{
          "type":"string"
        },
        "registered_address_town":{
          "type":"string"
        },
        "registered_address_county":{
          "type":"string"
        },
        "registered_address_postcode":{
          "type":"string"
        },
        "registered_address_country":{
          "type":"string",
          "format":"uuid"
        },
        "company_number":{
          "type":"string"
        },
        "alias":{
          "type":"string",
          "example":"Freds Bricks"
        },
        "business_type":{
          "type":"string",
          "format":"uuid",
          "example":"d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "sector":{
          "type":"string",
          "format":"uuid",
          "example":"d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "employee_range":{
          "type":"string",
          "format":"uuid",
          "example":"d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "turnover_range":{
          "type":"string",
          "format":"uuid",
          "example":"d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "account_manager":{
          "type":"string",
          "format":"uuid",
          "example":"d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "lead":{
          "type":"boolean"
        },
        "description":{
          "type":"string"
        },
        "website":{
          "type":"string",
          "format":"url",
          "example":"http://www.test.com"
        },
        "uk_region":{
          "type":"string",
          "format":"uuid",
          "example":"d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "trading_address_1":{
          "type":"string",
          "example":"1 Smith Street"
        },
        "trading_address_2":{
          "type":"string"
        },
        "trading_address_3":{
          "type":"string"
        },
        "trading_address_4":{
          "type":"string"
        },
        "trading_address_town":{
          "type":"string"
        },
        "trading_address_county":{
          "type":"string"
        },
        "trading_address_postcode":{
          "type":"string"
        },
        "trading_address_country":{
          "type":"string",
          "format":"uuid"
        }
      }
    },
    "CompaniesHouseCompany":{
      "type":"object",
      "required":[
        "company_number",
        "name"
      ],
      "properties":{
        "company_number":{
          "type":"string",
          "example":12387123978
        },
        "name":{
          "type":"string",
          "example":"Widgets r us"
        },
        "registered_address_1":{
          "type":"string",
          "example":"1 Smith Street"
        },
        "registered_address_2":{
          "type":"string"
        },
        "registered_address_3":{
          "type":"string"
        },
        "registered_address_4":{
          "type":"string"
        },
        "registered_address_town":{
          "type":"string"
        },
        "registered_address_county":{
          "type":"string"
        },
        "registered_address_postcode":{
          "type":"string"
        },
        "registered_address_country":{
          "type":"string",
          "format":"uuid"
        },
        "company_category":{
          "type":"string"
        },
        "company_status":{
          "type":"string"
        },
        "sic_code_1":{
          "type":"string"
        },
        "sic_code_2":{
          "type":"string"
        },
        "sic_code_3":{
          "type":"string"
        },
        "sic_code_4":{
          "type":"string"
        },
        "incorporation_date":{
          "type":"string",
          "format":"date"
        }
      }
    },
    "Contact":{
      "type":"object",
      "required":[
        "id",
        "company"
      ],
      "properties":{
        "id":{
          "type":"string",
          "format":"uuid",
          "example":"d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "company":{
          "type":"string",
          "format":"uuid",
          "example":"d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "title":{
          "type":"string",
          "format":"uuid",
          "example":"d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "first_name":{
          "type":"string",
          "example":"Widgets r us"
        },
        "last_name":{
          "type":"string",
          "example":"Widgets r us"
        },
        "job_title":{
          "type":"string",
          "example":"Widgets r us"
        },
        "advisor":{
          "type":"string",
          "format":"uuid",
          "example":"d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "primary":{
          "type":"boolean"
        },
        "telephone_nmber":{
          "type":"string"
        },
        "email":{
          "type":"string",
          "format":"email"
        },
        "address_same_as_company":{
          "type":"boolean"
        },
        "address_1":{
          "type":"string",
          "example":"1 Smith Street"
        },
        "address_2":{
          "type":"string"
        },
        "address_3":{
          "type":"string"
        },
        "address_4":{
          "type":"string"
        },
        "address_town":{
          "type":"string"
        },
        "address_county":{
          "type":"string"
        },
        "address_postcode":{
          "type":"string"
        },
        "address_country":{
          "type":"string",
          "format":"uuid"
        },
        "telephone_alternative":{
          "type":"string"
        },
        "email_alternative":{
          "type":"string",
          "format":"email"
        },
        "notes":{
          "type":"string"
        },
        "created_on":{
          "type":"string",
          "format":"date"
        }
      }
    },
    "Metadata":{
      "type":"object",
      "required":[
        "id",
        "name"
      ],
      "properties":{
        "id":{
          "type":"string",
          "format":"uuid",
          "example":"d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "name":{
          "type":"string"
        }
      }
    },
    "InvestmentProject":{
      "type":"object",
      "required":[
        "id",
        "company"
      ],
      "properties":{
        "id":{
          "type":"string",
          "format":"uuid",
          "example":"d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "company":{
          "type":"string",
          "format":"uuid",
          "example":"d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "name":{
          "type":"string",
          "example":"Invest in us"
        },
        "value":{
          "type":"string",
          "example":"High"
        },
        "state":{
          "type":"string",
          "example":"Prospect"
        },
        "land_date":{
          "type":"string",
          "format":"date"
        },
        "state_date":{
          "type":"string",
          "format":"date"
        },
        "open":{
          "type":"boolean"
        },
        "created_on":{
          "type":"string",
          "format":"date"
        }
      }
    },
    "SearchResults":{
      "type":"object",
      "properties":{
        "total":{
          "type":"integer"
        },
        "hits":{
          "type":"array",
          "items":{
            "$ref":"#/definitions/Hit"
          }
        }
      }
    },
    "Hit":{
      "type":"object",
      "required":[
        "_id",
        "_index",
        "_score",
        "_source"
      ],
      "properties":{
        "_id":{
          "type":"string"
        },
        "_index":{
          "type":"string"
        },
        "_score":{
          "type":"number",
          "format":"float"
        },
        "_source":{
          "$ref":"#/definitions/SearchResult"
        }
      }
    },
    "SearchResult":{
      "type":"object",
      "properties":{
        "id":{
          "type":"string",
          "format":"uuid",
          "example":"d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "name":{
          "type":"string",
          "example":"Widgets r us"
        },
        "registered_address_1":{
          "type":"string",
          "example":"1 Smith Street"
        },
        "registered_address_2":{
          "type":"string"
        },
        "registered_address_3":{
          "type":"string"
        },
        "registered_address_4":{
          "type":"string"
        },
        "registered_address_town":{
          "type":"string"
        },
        "registered_address_county":{
          "type":"string"
        },
        "registered_address_postcode":{
          "type":"string"
        },
        "registered_address_country":{
          "type":"string",
          "format":"uuid"
        },
        "company_number":{
          "type":"string"
        },
        "alias":{
          "type":"string",
          "example":"Freds Bricks"
        },
        "trading_address_1":{
          "type":"string",
          "example":"1 Smith Street"
        },
        "trading_address_2":{
          "type":"string"
        },
        "trading_address_3":{
          "type":"string"
        },
        "trading_address_4":{
          "type":"string"
        },
        "trading_address_town":{
          "type":"string"
        },
        "trading_address_county":{
          "type":"string"
        },
        "trading_address_postcode":{
          "type":"string"
        },
        "trading_address_country":{
          "type":"string",
          "format":"uuid"
        },
        "incorporation_date":{
          "type":"string",
          "format":"date"
        }
      }
    }
  },
  "host":"data-hub-zorg.herokuapp.com",
  "basePath":"https://data-hub-zorg.herokuapp.com/"
}